Factor out concern for mets to address class length
- Remove unused METs classes, inherited from another app. Can re-introduce if / when needed.
